Continue iterating on decoding support and memory management
This change continues to flesh out how KMIP structures are
decoded from a KMIP TTLV encoding, dynamically building out the
KMIP structure for certain primitive objects and KMIP attributes
as the encoding is processed. Corresponding initialization and
deallocation routines are included for these initial decoded
structures. Included is an attempt (mirroring OpenSSH) to
securely zero memory leveraging memset_s (if available) or a
volatile function pointer to a zeroing routine.