Finish the primary implementation of the BIO client library
This change completes the primary implementation of the OpenSSL
BIO client library for libkmip. The Create, Get, and Destroy
operations, as implemented in demo_create.c, demo_get.c, and
demo_destroy.c, can be supported using the high-, mid-, and
low-level client API, depending on the using application's needs.