Add a first cut on the library documentation
This change adds the first pass on content for the library
documentation. It is managed and built using Sphinx, like PyKMIP
and SLUGS. A doc-requirements.txt file is included to help devs
setup the right dependencies if they need them. A bunch of
additional detail is needed under the API section. Definitely a
work-in-progress.