Fix minor memory leak in library test suite
This change fixes a minor memory leak that was found in the
library test suite. The addition of the Credential linked list in
the library context involves dynamically allocating space for
the linked list root. This is only ever cleaned up when the
context is destroyed. The context was never officially destroyed
in the test_buffer_full_and_resize test, so the linked list root
memory was leaked every time that test was called. This has been
fixed with the context now officially destroyed in all cases.