Update the BIO client to null-terminate object UUID strings
This change updates the BIO client to ensure that UUID strings
returned by it are null terminated. By default, KMIP-encoded
strings are not null terminated. If strings from a KMIP request
or response message are returned verbatim, users may inadvertently
run off of the end of the string while using it. This change
prevents that case from happening.