Update the API docs to clarify the status code and target objects
This change updates the API function signatures, making it clear
that these functions are intended for use primarily with symmetric
key objects. Also add detail to the value of the status code
returned by these function to clarify when the value is a result
of a libkmip error vs a KMIP server error.