Add support for KMIP 2.0-style attributes
This change adds support for KMIP 2.0-style attributes, which
eschew the name/index/value tuple found in older versions of KMIP
for direct encodings of each value structure. This change forks
the attribute encode/decode functions, keying off of the KMIP
context's KMIP version value to determine how encoding/decoding
should be done.

Attribute-related utilities are also included, supporting the
inspection of the current context buffer for the next tag to
determine if it is a potential KMIP attribute.

Unit tests for all of these changes have been added to the unit
test suite.